coproc: Move script_context backend into new file
- No logic changes in this change, only moves all logic relevent to
creating materialized logs into a new file called script_context_backend.h/cc
